+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Tally.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
@@ -19,21 +20,10 @@
 module Cardano.Ledger.Conway.Rules.Tally (
   ConwayTALLY,
   TallyEnv (..),
-  GovernanceMetadata (..),
-  GovernanceProcedure (..),
   ConwayTallyPredFailure,
 ) where
 
-import Cardano.Ledger.BaseTypes (EpochNo (..), ShelleyBase, Url)
-import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
-import Cardano.Ledger.Binary.Coders (
-  Decode (..),
-  Encode (..),
-  decode,
-  encode,
-  (!>),
-  (<!),
- )
+import Cardano.Ledger.BaseTypes (EpochNo (..), ShelleyBase)
 import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Conway.Era (ConwayTALLY)
 import Cardano.Ledger.Conway.Governance (
@@ -41,15 +31,16 @@ import Cardano.Ledger.Conway.Governance (
   GovernanceAction,
   GovernanceActionId (..),
   GovernanceActionState (..),
-  VoteDecision,
+  GovernanceProcedure (..),
+  ProposalProcedure (..),
+  Vote,
   VoterRole,
+  VotingProcedure (..),
  )
-import Cardano.Ledger.Core (Era (..), EraPParams)
+import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential)
-import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Keys (KeyHash (..), KeyRole (..))
 import Cardano.Ledger.Rules.ValidationMode (Inject (..), Test, runTest)
-import Cardano.Ledger.SafeHash (SafeHash)
 import Cardano.Ledger.Shelley.Tx (TxId (..))
 import Control.DeepSeq (NFData)
 import Control.State.Transition.Extended (
@@ -58,7 +49,6 @@ import Control.State.Transition.Extended (
   TransitionRule,
   judgmentContext,
  )
-import Data.ByteString (ByteString)
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Sequence (Seq (..))
@@ -73,100 +63,14 @@ data TallyEnv era = TallyEnv
   , teRoles :: !(Map (Credential 'Voting (EraCrypto era)) VoterRole)
   }
 
-data GovernanceMetadata c = GovernanceMetadata
-  { gmUrl :: !Url
-  , gmHash :: !(SafeHash c ByteString)
-  }
-  deriving (Eq, Generic)
-
-instance Crypto c => DecCBOR (GovernanceMetadata c) where
-  decCBOR =
-    decode $
-      RecD GovernanceMetadata
-        <! From
-        <! From
-
-instance Crypto c => EncCBOR (GovernanceMetadata c) where
-  encCBOR GovernanceMetadata {..} =
-    encode $
-      Rec GovernanceMetadata
-        !> To gmUrl
-        !> To gmHash
-
-instance NoThunks (GovernanceMetadata era)
-
-instance NFData (GovernanceMetadata era)
-
-deriving instance Show (GovernanceMetadata era)
-
-data GovernanceProcedure era
-  = VotingProcedure
-      !(GovernanceActionId (EraCrypto era))
-      !VoterRole
-      !(Credential 'Voting (EraCrypto era))
-      !VoteDecision
-      !(GovernanceMetadata (EraCrypto era))
-  | ProposalProcedure
-      !Coin
-      !(KeyHash 'Staking (EraCrypto era))
-      !(GovernanceAction era)
-      !(GovernanceMetadata (EraCrypto era))
-  deriving (Eq, Generic)
-
-instance Era era => NFData (ConwayTallyPredFailure era)
-
-instance
-  ( Era era
-  , EraPParams era
-  ) =>
-  DecCBOR (GovernanceProcedure era)
-  where
-  decCBOR =
-    decode $
-      Summands "GovernanceProcedure" dec
-    where
-      dec 0 =
-        SumD VotingProcedure
-          <! From
-          <! From
-          <! From
-          <! From
-          <! From
-      dec 1 =
-        SumD ProposalProcedure
-          <! From
-          <! From
-          <! From
-          <! From
-      dec n = Invalid n
-
-instance (Era era, EraPParams era) => EncCBOR (GovernanceProcedure era) where
-  encCBOR (VotingProcedure gaid role cred decision gmd) =
-    encode @_ @(GovernanceProcedure era) $
-      Sum VotingProcedure 0
-        !> To gaid
-        !> To role
-        !> To cred
-        !> To decision
-        !> To gmd
-  encCBOR (ProposalProcedure c kh ga gmd) =
-    encode @_ @(GovernanceProcedure era) $
-      Sum ProposalProcedure 1
-        !> To c
-        !> To kh
-        !> To ga
-        !> To gmd
-
-instance EraPParams era => NoThunks (GovernanceProcedure era)
-
-instance EraPParams era => NFData (GovernanceProcedure era)
-
-deriving instance EraPParams era => Show (GovernanceProcedure era)
-
 data ConwayTallyPredFailure era
   = VoterDoesNotHaveRole !(Credential 'Voting (EraCrypto era)) !VoterRole
   | GovernanceActionDoesNotExist !(GovernanceActionId (EraCrypto era))
   deriving (Eq, Show, Generic)
+
+instance Era era => NFData (ConwayTallyPredFailure era)
+
+instance Era era => NoThunks (ConwayTallyPredFailure era)
 
 instance Era era => STS (ConwayTALLY era) where
   type State (ConwayTALLY era) = ConwayTallyState era
@@ -184,7 +88,7 @@ addVote ::
   GovernanceActionId (EraCrypto era) ->
   VoterRole ->
   Credential 'Voting (EraCrypto era) ->
-  VoteDecision ->
+  Vote ->
   ConwayTallyState era ->
   ConwayTallyState era
 addVote gaid role kh decision (ConwayTallyState st) =
@@ -243,19 +147,21 @@ tallyTransition = do
   let updateState st' Empty = pure st'
       updateState !st' ((x, idx) :<| xs) = do
         st'' <- case x of
-          VotingProcedure gaid role kh vote _ -> do
-            runTest $ noSuchGovernanceAction st gaid
-            runTest $ voterHasRole @era kh role vRoles
-            pure $ addVote gaid role kh vote st'
-          ProposalProcedure c kh action _ ->
-            pure $
-              addAction
-                epoch
-                (GovernanceActionId txid idx)
-                c
-                kh
-                action
-                st'
+          GovernanceVotingProcedure
+            VotingProcedure {vProcGovActionId, vProcRole, vProcRoleKeyHash, vProcVote} -> do
+              runTest $ noSuchGovernanceAction st vProcGovActionId
+              runTest $ voterHasRole @era vProcRoleKeyHash vProcRole vRoles
+              pure $ addVote vProcGovActionId vProcRole vProcRoleKeyHash vProcVote st'
+          GovernanceProposalProcedure
+            ProposalProcedure {pProcDeposit, pProcReturnAddr, pProcGovernanceAction} ->
+              pure $
+                addAction
+                  epoch
+                  (GovernanceActionId txid idx)
+                  pProcDeposit
+                  pProcReturnAddr
+                  pProcGovernanceAction
+                  st'
         updateState st'' xs
 
   updateState st $ Seq.zip govProcedures [0 ..]