+++ eras/conway/test-suite/test/Test/Cardano/Ledger/Conway/Serialisation/CDDL.hs
@@ -12,7 +12,7 @@ import Cardano.Ledger.Alonzo.Tx (Data)
 import Cardano.Ledger.Alonzo.TxWits (Redeemers)
 import Cardano.Ledger.Babbage.TxBody (Datum)
 import Cardano.Ledger.Conway (Conway)
-import Cardano.Ledger.Conway.Rules (GovernanceProcedure)
+import Cardano.Ledger.Conway.Governance (ProposalProcedure, VotingProcedure)
 import Cardano.Ledger.Core
 import qualified Data.ByteString.Lazy as BSL
 import Paths_cardano_ledger_conway_test (getDataFileName)
@@ -39,7 +39,8 @@ tests n = withResource combinedCDDL (const (pure ())) $ \cddl ->
     , cddlTest @CostModels (eraProtVerHigh @Conway) n "costmdls"
     , cddlAnnotatorTest @(Redeemers Conway) (eraProtVerHigh @Conway) n "[* redeemer]"
     , cddlAnnotatorTest @(Tx Conway) (eraProtVerHigh @Conway) n "transaction"
-    , cddlTest @(GovernanceProcedure Conway) (eraProtVerHigh @Conway) n "governance_procedure"
+    , cddlTest @(VotingProcedure Conway) (eraProtVerHigh @Conway) n "voting_procedure"
+    , cddlTest @(ProposalProcedure Conway) (eraProtVerHigh @Conway) n "proposal_procedure"
     ]
       <*> pure cddl