+++ libs/cardano-ledger-pretty/src/Cardano/Ledger/Pretty/Conway.hs
@@ -20,21 +20,23 @@ import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Delegation.Certificates (ConwayDCert (..), toShelleyDCert)
 import Cardano.Ledger.Conway.Governance (
+  Anchor,
   ConwayGovernance (..),
   ConwayTallyState (..),
   GovernanceAction (..),
   GovernanceActionId (..),
   GovernanceActionIx (..),
   GovernanceActionState (..),
-  VoteDecision (..),
+  GovernanceProcedure,
+  ProposalProcedure (..),
+  Vote (..),
   VoterRole (..),
+  VotingProcedure (..),
  )
 import Cardano.Ledger.Conway.Rules (
   ConwayLedgerPredFailure (..),
   ConwayTallyPredFailure,
   EnactState (..),
-  GovernanceMetadata,
-  GovernanceProcedure,
   PredicateFailure,
   RatifyState (..),
  )
@@ -105,10 +107,17 @@ ppConwayTxBody txb =
     , ("script integrity hash", ppStrictMaybe ppSafeHash $ txb ^. scriptIntegrityHashTxBodyL)
     , ("auxiliary data hash", ppStrictMaybe ppAuxiliaryDataHash $ txb ^. auxDataHashTxBodyL)
     , ("network id", ppStrictMaybe ppNetwork $ txb ^. networkIdTxBodyL)
-    , ("governance procedures", ppStrictSeq prettyA $ txb ^. govProcsTxBodyL)
+    , ("voting procedures", ppStrictSeq prettyA $ txb ^. votingProceduresTxBodyL)
+    , ("proposal procedures", ppStrictSeq prettyA $ txb ^. proposalProceduresTxBodyL)
     ]
 
-instance PrettyA VoteDecision where
+instance PrettyA Vote where
+  prettyA = viaShow
+
+instance EraPParams era => PrettyA (VotingProcedure era) where
+  prettyA = viaShow
+
+instance EraPParams era => PrettyA (ProposalProcedure era) where
   prettyA = viaShow
 
 instance EraPParams era => PrettyA (GovernanceProcedure era) where
@@ -181,7 +190,7 @@ instance PrettyA GovernanceActionIx where
 instance PrettyA VoterRole where
   prettyA = ppString . show
 
-instance PrettyA (GovernanceMetadata era) where
+instance PrettyA (Anchor era) where
   prettyA = viaShow
 
 instance PrettyA (PParamsUpdate era) => PrettyA (GovernanceActionState era) where