+++ eras/conway/impl/CHANGELOG.md
@@ -25,6 +25,10 @@
 * Add `Anchor` and `AnchorDataHash` types. #3323
 * Rename `transDCert` to `toShelleyDCert`
 * Add `fromShelleyDCertMaybe`
+* Renamed `Vote` type to `VotingProcedure`
+* Add `ProposalProcedure`
+* Use `VotingProcedure` and `ProposalProcedure` in `GovernanceProcedure`
+* Rename `VoteDecision` to `Vote`. Rename `No`/`Yes` -> `VoteNo`/`VoteYes`.
 
 ### `testlib`